requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic[email]
pydantic-settings>=2.0.0  # For settings management

# HTTP & API
requests>=2.31.0
aiohttp>=3.9.0  # Async HTTP client for live scores
websockets>=12.0  # WebSocket support for FastAPI
pandas>=2.1.0  # Data analysis for MoneyPuck CSV data
httpx>=0.25.0  # Modern HTTP client

# Environment & Configuration
python-dotenv>=1.0.0

# AI/ML
openai>=1.3.0

# Database (PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Automation & Scheduling
schedule>=1.2.0  # Task scheduling
apscheduler>=3.10.0  # Advanced job scheduling

# Caching
cachetools>=5.3.0  # Caching utilities

# Monitoring & Logging
prometheus-client>=0.19.0  # Metrics collection
python-json-logger>=2.0.7  # Structured JSON logging

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=20.0.0  # Generate fake data for tests

# Code Quality
black>=23.11.0  # Code formatting
flake8>=6.1.0  # Linting
mypy>=1.7.0  # Type checking
isort>=5.12.0  # Import sorting

# Documentation
mkdocs>=1.5.0  # Documentation generation
mkdocs-material>=9.4.0  # Material theme

# Optional: Production extras
gunicorn>=21.2.0  # Production WSGI server
